requirements.txt

numpy
numexpr

Pillow
mercantile
boto3

rasterio[s3]>=1.0a12
rio-toa
rio-pansharpen

shapely